Sample visualizer updates and bugfixes
* Update reflectivity visualization for the upcoming 2.1 firmware
release
* Fix a bug causing the viewport to be sized incorrectly on retina
displays
* Some code reorganization in the sample visualizer library
