+++ cmake/Findjsoncpp.cmake
@@ -1,12 +1,19 @@
-# Try to find a cmake config compatible with vcpkg and fall back to defining
-# targets using pkgconfig, which seems more consistent for jsoncpp across
-# different platforms.
+# Targets defined for jsoncpp differ a good amount between distributions
 
 include(FindPackageHandleStandardArgs)
 
-# recent jsoncpp config shipped with vcpkg defines this target
 find_package(jsoncpp CONFIG QUIET)
-if(WIN32 OR (jsoncpp_FOUND AND TARGET jsoncpp_lib))
+
+# This target exists on ubuntu / debian. For some reason debian bullseye doesn't
+# ship a static lib / define the jsoncpp_lib_static target.
+if(TARGET jsoncpp_lib)
+  find_package_handle_standard_args(jsoncpp DEFAULT_MSG jsoncpp_CONFIG)
+  return()
+# Target available in recent versions of vcpkg. We can make an alias of an
+# imported target since we're guaranteed to be using a recent cmake
+elseif(TARGET jsoncpp_static)
+  set_target_properties(jsoncpp_static PROPERTIES IMPORTED_GLOBAL TRUE)
+  add_library(jsoncpp_lib ALIAS jsoncpp_static)
   find_package_handle_standard_args(jsoncpp DEFAULT_MSG jsoncpp_CONFIG)
   return()
 endif()
