+++ ouster_client/src/netcompat.cpp
@@ -33,21 +33,11 @@ static StaticWrapper resources = {};
 #endif
 
 int socket_close(SOCKET sock) {
-    int status = 0;
-
 #ifdef _WIN32
-    status = shutdown(sock, SD_BOTH);
-    if (status == 0) {
-        status = closesocket(sock);
-    }
+    return closesocket(sock);
 #else
-    status = shutdown(sock, SHUT_RDWR);
-    if (status == 0) {
-        status = close(sock);
-    }
+    return close(sock);
 #endif
-
-    return status;
 }
 
 std::string socket_get_error() {
