+++ ouster_viz/CMakeLists.txt
@@ -11,12 +11,13 @@ find_package(GLEW REQUIRED)
 find_package(glfw3 CONFIG REQUIRED)
 
 # ==== Libraries ====
-add_library(ouster_viz src/point_viz.cpp src/cloud.cpp src/camera.cpp src/image.cpp)
+add_library(ouster_viz src/point_viz.cpp src/cloud.cpp src/camera.cpp src/image.cpp
+  src/image_processing.cpp src/lidar_scan_viz.cpp)
 target_link_libraries(ouster_viz
   PUBLIC Threads::Threads glfw GLEW::GLEW OpenGL::GL
   PRIVATE ouster_client)
 target_include_directories(ouster_viz PUBLIC include)
 
 # ==== Executables ====
-add_executable(simple_viz src/main.cpp)
+add_executable(simple_viz src/simple_viz.cpp)
 target_link_libraries(simple_viz ouster_client ouster_viz)
