+++ ouster_viz/src/image.cpp
@@ -1,3 +1,6 @@
+#include "image.h"
+
+#include <GL/glew.h>
 
 #include "ouster/point_viz.h"
